# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Stock Advisory + MLM Backend
  version: 0.1.0
servers:
  - url: http://localhost:8080
  - url: https://juststock-server-q6c9.onrender.com
paths:
  /health:
    get:
      summary: Healthcheck
      responses:
        '200': { description: OK }

  /auth/signup:
    post:
      summary: Signup with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                confirmPassword: { type: string, minLength: 8 }
                mobile: { type: string }
                referralCode: { type: string, description: Optional referrer code }
              required: [name, email, password, confirmPassword, mobile]
      responses:
        '201': { description: Tokens }

  /auth/login:
    post:
      summary: Login with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                password: { type: string }
              required: [email, password]
      responses:
        '200': { description: Tokens }

  /auth/refresh:
    post:
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string }
              required: [refreshToken]
      responses:
        '200': { description: Access token }

  /auth/request-otp:
    post:
      summary: Request login OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: { email: { type: string, format: email } }
              required: [email]
      responses:
        '200': { description: Sent }

  /auth/verify-otp:
    post:
      summary: Verify OTP and issue tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                code: { type: string }
              required: [email, code]
      responses:
        '200': { description: Tokens }

  /auth/forgot-password:
    post:
      summary: Send password reset OTP to email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: { email: { type: string, format: email } }
              required: [email]
      responses:
        '200': { description: Sent }

  /auth/reset-password:
    post:
      summary: Reset password using OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                code: { type: string }
                newPassword: { type: string, minLength: 8 }
                confirmPassword: { type: string, minLength: 8 }
              required: [email, code, newPassword, confirmPassword]
      responses:
        '200': { description: OK }

  /calls:
    get:
      summary: List calls; body only when unlocked
      responses:
        '200': { description: OK }

  /calls/latest5:
    get:
      summary: Latest one message per category (NIFTY, BANKNIFTY, SENSEX, COMMODITY, OTHERS)
      responses:
        '200': { description: OK }

  /messages/latest5:
    get:
      summary: Alias of /calls/latest5
      responses:
        '200': { description: OK }

  /messages/latest/{category}:
    get:
      summary: Latest message in a category (body only if unlocked)
      parameters:
        - in: path
          name: category
          required: true
          schema: { type: string, enum: [NIFTY, BANKNIFTY, SENSEX, COMMODITY, OTHERS] }
      responses:
        '200': { description: OK }

  /messages/latest/{category}/unlock:
    post:
      summary: Unlock latest message in a category by paying from wallet
      parameters:
        - in: path
          name: category
          required: true
          schema: { type: string, enum: [NIFTY, BANKNIFTY, SENSEX, COMMODITY, OTHERS] }
      responses:
        '200': { description: Unlocked }
        '402': { description: Insufficient funds }

  /calls/{callId}:
    get:
      summary: Get a call if unlocked
      parameters:
        - in: path
          name: callId
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }

  /calls/{callId}/unlock:
    post:
      summary: Unlock a call
      parameters:
        - in: path
          name: callId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Unlocked }
        '402': { description: Insufficient funds }

  /admin/messages/{category}:
    post:
      summary: Admin create latest message for a category
      parameters:
        - in: path
          name: category
          required: true
          schema: { type: string, enum: [NIFTY, BANKNIFTY, SENSEX, COMMODITY, OTHERS] }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                body: { type: string }
                price: { type: integer, minimum: 0, default: 116 }
              required: [title, body]
      responses:
        '200': { description: Created }

  /users/me:
    get:
      summary: Get current user profile
      responses:
        '200': { description: OK }
    patch:
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                phone: { type: string }
      responses:
        '200': { description: OK }

  /users/me/wallet:
    get:
      summary: Get wallet balances
      responses:
        '200': { description: OK }

  /payments/registration/order:
    post:
      summary: Create registration order (₹2100)
      responses:
        '200': { description: Order }

  /payments/renewal/order:
    post:
      summary: Create renewal order (₹1000)
      responses:
        '200': { description: Order }

  /payments/webhook:
    post:
      summary: Razorpay webhook capture
      responses:
        '200': { description: OK }

  /wallet/topup/order:
    post:
      summary: Create wallet topup order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount: { type: integer, minimum: 1 }
              required: [amount]
      responses:
        '200': { description: Order }

  /wallet/topup/simple:
    post:
      summary: Simplified wallet topup (2100 for first, then 1000)
      responses:
        '200': { description: OK }

  /me/history/wallet:
    get:
      summary: Wallet transaction history
      responses:
        '200': { description: OK }

  /history/calls:
    get:
      summary: Unlocked calls history
      responses:
        '200': { description: OK }

  /users/me/history/commissions:
    get:
      summary: Commissions history
      responses:
        '200': { description: OK }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT